Add patch root-5.34.01-rev-45321-45322.patch
This is a new version of Matevz's patch to fix the I/O regression with ROOT
5.34.01.  The major difference is that this version of the patch has been
committed to the ROOT 5.34 branch.  Different implementation, same overall
effect.

This restores the semantics of 5.32.00 for TFile::SetCacheRead(0) if a
non-default flag is passed; the TFile will no longer use the prior cache,
but the cache can continue to use the TFile.

Request: https://hypernews.cern.ch/HyperNews/CMS/get/ext-pkg-mgmt/319.html