classlib: add guards for x86_64 unwind implementation
On ARMv7 architecture we are compiling x86_64 unwind implementation
which depends on libgcc_s implemenation. In particular _Unwind_GetIP
call, which is exported by libgcc_s library for x86_84, but on ARMv7
_Unwind_GetIP is a macro. Macro depends on _Unwind_GetGR, which is
not a public symbol in libgcc_s. Here we might be using private and
not cross-platform solution for unwind. ``classlib'' is no more
maintaind project.

The patch adds guards for x86_64 specific implemention and it's not
compiled on ARMv7 architecture anymore.

Signed-off-by: David Abdurachmanov <davidlt@cern.ch>