ncurses: disable terminfo, enable termcap as a fallback
ncurses is not designed to build w/o terminfo database. Disable
database, but enable termcap as a fallback mechanism. A fallback
is required. By default terminfo database is full of hard links,
which are not supported by AFS. If we upload a package with hard
links we cannot install it on AFS.

Signed-off-by: David Abdurachmanov <davidlt@cern.ch>