root: move from configure/Makefile to CMake
The patch rewrites build recipie to CMake instead of old
configure/Makefile. CMake is easier to manage to package mantainers and
CMake is widely used by ROOT team in their CI system. ROOT is not tested
by configure/Makefile.

A single patch is needed (to be upstreamed), because we are building on
OSX without Cocoa, but with X11. This is not a standard configuration
tested in ROOT CI system.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>
Tested-on: slc6_amd64_gcc491
Tested-on: fc19_aarch64_gcc490
Tested-on: osx10A_amd64_clang610