gdb: bump to 7.9.1
Older version of GDB contains bug which heavily slows down GDB
if LLVM symbols are involved. 7.9.1 provides huge speedup debugging
CMSSW (which incl. ROOT6 + LLVM).

An extra patch is required for `configure` to disable `makeinfo` if
such does not exist on the system. The patch brings `configure` to
it's documented behavior. There are BZ items for `MAKEINFO=true` not
working as advertised.

`configure` is patched instead of `configure.ac` because for
regeneration a very specific version of `autoconf` would be required,
which is too old on the system and too new in `autotools` package.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>
(cherry picked from commit 982af1c7054f2825ebd8f1f7b1f5ba70d18a04db)