openssl,openssl-bootstrap: keep only 1.0.2d; fix CA bundle
Use 1.0.2d even for SLC6. In newer glibc versions __secure_getenv was
renamed to secure_getenv, thus we just rename it back on SLC6.

Include a fix for PR3979.

Set /etc/pki/tls as openssldir, but make sure that installation step
does not touch the directory. This solves the problem where das_client.py
was not able to verify cmsweb certificate.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>