tbb: bump to 44_20151115oss; enable Clang C++11 features on Linux
Currently we cannot use use C++11 features provided by Clang on Linux
with TBB because of GCC version check as Clang always reports GCC 4.2.1
compability version. In addition to that we should not require libc++ on
Linux. On Linux distributions Clang is mostly used with GNU libstdc++
for binary compatibility.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>