nss-bootstrap: bring back libsoftokn3* libfreebl3* libsql*
Building these libaries in non-API or/and non-ABI compatible way can
break the system. Glibc depends on these libraries. Originally those
libraries were not packed in order to avoid breaking the system.

If we don't include them and NSS is updated on system we might have
broken NSS on our side. This what happened on Fedora 22 where NSS was
updated multiple times and finally RPM in our bootstrap driver was failing
once trying to load libsoftokn3* from the system, but it was not
compatible.

Thus add the libraries back.

Signed-off-by: David Abdurachmanov <davidlt@cern.ch>