qt,qt-toolfile: update to 4.8.7
Update to the lastest available version, but note that QT 4 is no more
supported product. Remove obsolete patch for GCC 4.7 and add a new one
for GCC 6 support.

Disable AVX and RPATHs.

JavaScript plugins are not building under GCC 6, thus disable
JavaScript/Scripting, which does not seem to be used in CMSSW.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>