nss-bootstrap: include nss-pem
This is included by default on Fedora and produces libnsspem.so, which
we lacked in our NSS builds. Before this change libnsspem.so was being
used from the system, but it might not be ABI/API compatible with our
NSS build.

This could break GIT trying to use HTTPS, e.g.:

    Peer's certificate issuer has been marked as not trusted by the user.

Signed-off-by: David Abdurachmanov <davidlt@cern.ch>