Removal of ncurses
CMSSW always required having ncurses installed on the system, the
exception is only ncurses-devel package, which is required on build
machines only. It's hard to image system without ncurses installed.

This should resolve issue two termcap/terminfo available in the
distribution which could be out-of-sync or/and  not compatible.

This could be visible while launching Python:

    [dlange@cmsdev01 src]$ python
    Python 2.7.11 (default, Jan 21 2016, 17:19:47)
    [GCC 5.3.0] on linux2
    Type "help", "copyright", "credits" or "license" for more information.
    Name collision 'vt52' between
    vt52|dec vt52
    and vt52|dec vt52
    ...now dec vt52
    Cannot remove alias 'dec vt52'
    Name collision 'vt100-nav' between
    vt100-nav|vt100 without advanced video option
    and vt100-nav|vt100 without advanced video option
    [..]

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>