icc-gcc-toolfile: tune ISA extensions
We need `-msse3` for DataFormats/Math otherwise we loose `hadd` function,
which is enabled starting SSE3.

Let ICC generate multiple dynamic paths in the binary for SSE4.1, SSE4.2,
AVX and AVX2. Note, CORE-AVX2 should be enough to enable all of this. We
might want to shorten this in the future.

Set base line as SSE3, i.e. if CPU does support higher features use SSE3
implementation as default.

Signed-off-by: David Abdurachmanov <david.abdurachmanov@gmail.com>