Update PHEDEX-lifecycle.spec for 1.3.0pre1
First try to accommodate new directory structure under Testbed/LifeCycle, migration to git, and more recent fixes. 
This pre-release does not include PHEDEX::Core, as it comes with PHEDEX dependency. 
We also do not  install the T0 component: to be confirmed with Dirk if it is still needed. If not, we may also skip the respective Provides: directives.