python,readline,sqlite: remove readline from CMSDIST
Within *-driver.txt we always requested the system to have readline.
We found that double reverse delete was segfaulting Python interpreter.
This was tracked down to a bug in readline, which was fixed in Debian
and Fedora. The fix most likely is part of readline 7.0.

We remove readline and use the one provided by the system.

Must have readline-devel package installed.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>