py2-psutil: new Python package
psutil (process and system utilities) is a cross-platform library for
retrieving information on running processes and system utilization (CPU,
memory, disks, network) in Python. It is useful mainly for system
monitoring, profiling and limiting process resources and management
of running processes.

Supports Python 2.6 - 3.5 and variety of operating systems (highly
portable).

Note: it's not a pure Python package, it does build shared library.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>