Disable automatic package dependencies
Llvm 7 includes a new tool, llvm-exegesis, that depepnds on libpfm.so.
Rather than including this library in the CMSSW stack, we assume that it
will be available form the host system; as it's not something strictly
necessary to have, we avoid a hard dependency on it in the RPM package.