Bump fastHadd version again
The 2.3 version was most likely build with a ROOT6.12 version that was missing the ROOT-9173 fix and therefore has the wrong `TH1` `ClassDef`. It must not be used in P5.

By now we should have that worked out, and we might need to switch to 2.4 in online once HLT switches to CMSSW10_2.