Add libpciaccess, a generic PCI access library
libpciaccess is a generic PCI access library originally developed for the
Xorg project; it is now used by various other projects, for example the
hwloc library developed by the OpenMPI project.

The upstream git repository is at
https://gitlab.freedesktop.org/xorg/lib/libpciaccess

To avoid the dependency on the "xorg-macros" autotool package, the CMSDIST
package for libpciaccess is based on the Debian package (where the autogen
step has already been performed) instead of the upstream git sources.