Disable aggressive compiler optimizations for VecGeom
It seems that gcc10 miscompiles (some of) the VecGeom code at -O3. In
my tests, switching back to -O2 fixes the crashes. More details and
discussion if that is indeed a compiler problem or an issue in the code
in https://sft.its.cern.ch/jira/projects/VECGEOM/issues/VECGEOM-600