Build alpaka code in GPU-only mode
Build alpaka device code for the CUDA and ROCm back-ends in "GPU only" mode.
In this mode, functions marked as ALPAKA_FN_ACC are marked as __device__
functions, and are compiled only for the corresponding GPU device back-ends.

Currently, functions marked as ALPAKA_FN_ACC are marked as __host__ __device__
functions, and may be compiled for both device and host back-ends. The latter
leads to linker errors in kernels that use device symbols like threadIdx,
blockIdx, etc. that are compiled for the ROCm back-end.