RDMA Core: make libibverbs plugins relocatable
Allow overriding the libibverbs configuration directory setting the
environment variable VERBS_CONFIG_DIR.

Set VERBS_CONFIG_DIR in the scram tool to point to the libibverbs
configuration directory.

Include the libibverbs configuration directory in the package. Update
the libibverbs configuration to include the full path to the plugins,
and relocate them in the post-install section.

Include the user binaries in the package.