Update alpaka to 1.3.0
This is primarily a bugfix release with a focus on SYCL changes.

It addresses several SYCL-specific issues, including fixes to the parallel loop
patterns for the Intel FPGA SYCL backend, corrections to the SYCL index order,
and updates to template arguments of SYCL buffer specializations.

Additionally, the release includes changes to SYCL attributes and protections
for __SYCL_TARGET macros.

Overall, this release aims to improve stability and performance, particularly
for SYCL-related functionalities.